Fix to the thread timing issue.
It seems to have been a waiting issue with the tasks not fully having time to be started up before the process was checking them.  It's quite suspicious in general how that was working, but introducing a simple 500ms moratorium is enough to dodge the issue on current and all future machines.